refactor(sampling): separate executor responsibilities
Move CPU plan lowering and multi-shot result orchestration out of the mutable one-shot executor. Encapsulate optional ISA-specific fused-rotation preparation behind a portable dispatch object while preserving the existing action tape and scalar fallback.
