+++ src/clifft/api/statevector.cc
@@ -1,4 +1,5 @@
 #include "clifft/sampling/executor.h"
+#include "clifft/sampling/state_queries.h"
 #include "clifft/svm/svm.h"
 #include "clifft/svm/svm_math.h"
