+++ src/python/bindings.cc
@@ -20,8 +20,9 @@
 #include "clifft/optimizer/statevector_squeeze_pass.h"
 #include "clifft/optimizer/swap_meas_pass.h"
 #include "clifft/optimizer/tile_axis_fusion_pass.h"
-#include "clifft/sampling/executor.h"
 #include "clifft/sampling/planner.h"
+#include "clifft/sampling/sampler.h"
+#include "clifft/sampling/state_queries.h"
 #include "clifft/svm/svm.h"
 #include "clifft/util/config.h"
 #include "clifft/util/introspection.h"
