+++ tests/test_importance_sampling.cc
@@ -4,8 +4,8 @@
 #include "clifft/backend/backend.h"
 #include "clifft/circuit/parser.h"
 #include "clifft/frontend/frontend.h"
-#include "clifft/sampling/executor.h"
 #include "clifft/sampling/planner.h"
+#include "clifft/sampling/sampler.h"
 #include "clifft/svm/svm.h"
 #include "clifft/util/fault_sampling.h"
 #include "clifft/util/xoshiro.h"
